Fix Linux build issue

+++ src/plugins/intel_gpu/src/graph/impls/ocl_v2/group_normalization_fsv16_fused.hpp
@@ -58,7 +58,7 @@ struct GroupNormalizationFsv16Fused : public GroupNormalizationBase {
             return false;
         }
 
-        constexpr size_t fsv = 16;
+        constexpr int64_t fsv = 16;
         // feature count needs to be static for following checks
         if (in0_layout.get_partial_shape()[1].is_dynamic())
             return false;
